requirements.txt
PySide6
pyinstaller
Pillow
pyenchant